Remove spec validation tests and disable others that need refactoring
