+++ tests/resource_response_test.py
@@ -61,11 +61,12 @@ from mock import Mock
 
 import httpretty
 from dateutil.tz import tzutc
+import pytest
 from requests import HTTPError
 
 from bravado.client import SwaggerClient
 from bravado.exception import CancelledError
-from exception import SwaggerError
+from bravado.exception import SwaggerError
 from bravado.response import HTTPFuture
 
 
@@ -88,6 +89,7 @@ class HTTPFutureTest(unittest.TestCase):
         self.assertFalse(self.future.cancelled())
 
 
+@pytest.mark.xfail(reason='Re-write when Operation ported to Swagger 2.0')
 class ResourceResponseTest(unittest.TestCase):
     def setUp(self):
         parameter = {
@@ -252,11 +254,6 @@ class ResourceResponseTest(unittest.TestCase):
         future = SwaggerClient.from_url(
             u'http://localhost/api-docs').api_test.testHTTP(test_param="a")
         self.assertTrue(isinstance(future, HTTPFuture))
-
-    # TODO test timeout : delay/sleep in httpretty body doesn't work.
-    @httpretty.activate
-    def test_timeout_works_for_sync_http_client(self):
-        pass
 
 
 if __name__ == '__main__':
